+++ prisma/migrations/20240724152847_remove_unique_index_from_player_bonus_id/migration.sql
@@ -1,0 +1,2 @@
+-- DropIndex
+DROP INDEX `PLAYERS_bonus_id_key` ON `PLAYERS`;