+++ prisma/migrations/20240725141827_remove_bonus_id_from_player/migration.sql
@@ -1,0 +1,8 @@
+/*
+  Warnings:
+
+  - You are about to drop the column `bonus_id` on the `PLAYERS` table. All the data in the column will be lost.
+
+*/
+-- AlterTable
+ALTER TABLE `PLAYERS` DROP COLUMN `bonus_id`;