+++ src/types/request/bonus.ts
@@ -1,0 +1,14 @@
+export type BonusSettings = {
+  percentage: number;
+  amount: number;
+};
+
+export type CreateBonusProps = BonusSettings & {
+  player_id: string;
+  status?: string;
+};
+
+export type BonusUpdatableProps = {
+  amount?: number;
+  status: string;
+};