+++ src/types/response/bonus.ts
@@ -1,0 +1,7 @@
+import { Bonus } from "@prisma/client";
+
+export type BonusRedemptionResult = {
+  player_balance?: number;
+  error?: string;
+  bonus: Bonus;
+};